Refactor to use behaviour framework. Python 3 compatibility.
