+++ commands.ini
@@ -1,37 +1,9 @@
 [Commands]
-(what time is it|what is the time|time)\??$:general.time
-list commands|help|command list:general.command_list
-weather forecast:weather.forecast
-weather:weather.get
-word of the day:word_of_the_day.get
-check fibre:fibre_checker.check
-update:general.update_self
-shutdown pi:general.shutdown_self
-reboot pi:general.reboot_self
-news sources:news.news_api_sources
-news:news.top_stories
-morning others$:general.morning_others
-morning|good morning$:general.morning
 camera:camera.take_photo
 video:camera.take_video
-get log:general.get_log
-rotate log:general.rotate_log
-^[$-1234567890.]* [a-zA-Z\s]*:expenses.log_expense
-budget|how much money:expenses.expenses_remaining_weekly
-get transactions:monzo.log_all_transactions
-get recent transactions:monzo.log_recent_transactions
-add monzo token:monzo.add_token
-get expenses:expenses.expenses_get
-broadcast:general.broadcast
-riddle answer:riddle.answer
-riddle:riddle.get_next
 goodbye|bye|we're going out|security on:security.on
 hello|we're (back|home|here)|security off:security.off
 test security:security.test
 is house empty:security.house_empty
 sweep:security.sweep
 override security:security.override
-thought of the day:reddit.shower_thought
-did you know|teach me something:reddit.did_you_know
-new houses:zoopla.get_properties
-houses:zoopla.get_all_properties
