--- files/riddle_test.htm
@@ -1,4 +1,0 @@
-<li>It is greater than God and more evil than the devil. The poor have it, the rich need it and if you eat it you'll die. What is it?<br />
-<em>Nothing.</em></li>
-<li>It walks on four legs in the morning, two legs at noon and three legs in the evening. What is it?<br />
-<em>Man</em></li>
